[misc] Refactor plugin as part of larger renovations
This is the beginning work on freshening up this plugin a bit.

1) Refactors helper functions into separate modules for reuse.
2) Cleanly delineates between DD error handling and helper error handling.
3) Adds in logging for when DD returns a non 2xx code.  (Thanks @directionless!)
