+++ lib/lita-datadog.rb
@@ -4,6 +4,8 @@ Lita.load_locales Dir[File.expand_path(
   File.join('..', '..', 'locales', '*.yml'), __FILE__
 )]
 
+require 'lita/helpers/utilities'
+require 'lita/helpers/graphs'
 require 'lita/handlers/datadog'
 
 require 'dogapi'
